requirements.txt
aiohttp
discord.py
pre-commit
jishaku
python-dotenv